Update depratio_sco.csv
Updated figures for Scotland dependency ratio as given at http://statistics.gov.scot/data/population-estimates-dependency
Note: This indicator is an estimate of the dependency ratio - children and older people (aged 65 and over) expressed as a percentage of people aged 16 to 64. Please note the denominator for this indicator has been changed from the former working age population (16-59 years for women and 16-64 years for men) to population aged 16-64 years for both men and women. Time series on the new 16-64 basis should now be used to ensure consistent comparisons over time.